MTE and MTR sub-estimation functions

These functions allow to estimate the mte and mtr, and their confidence intervals, based on coefficients estimated from the main model in the main function. More details can be found in Bruneel-Zupanc (2024).
Different formulas must be applied depending on whether the treatment is homogenous or heterogenous.

Usage

mtr_fun(coeff, vcov, var_treatment, var_cov_2nd, ref_indiv, seq_u,
               homogenous=FALSE, pol_degree, conf_level, t_value, Xdat)