MTE and MTR sub-estimation functions

These functions allow to estimate the mte and mtr, and their confidence intervals, based on coefficients estimated from the main model in the main function. More details can be found in Bruneel-Zupanc (2024).
Different formulas must be applied depending on whether the treatment is homogenous or heterogenous.

Usage

mtr_fun = function(d, data, ref_indiv, seq_u,
                         bwd = NULL, bw_y = NULL, bw_method = "plug-in",
                         pol_degree1, pol_degree2, var_outcome, var_treatment, var_w0, var_w1, var_covariates)

lpoly_regress(x, y, bw=NULL, bw_method="plug-in", degree, drv)

mtr_fun_sieve(coeff, vcov, var_treatment, var_cov_2nd, ref_indiv, seq_u,
               homogenous=FALSE, pol_degree, conf_level, t_value, Xdat)

Arguments

bw_method

Method to compute the bandwidth (if bandwdith is NULL)

x

Vector of x values

y

Vector of y values

bw

Pre-specified bandwidth

degree

Degree of the polynomial: recommended to set to drv + 1

drv

Derivative order of the function to be estimated.