MTE and MTR sub-estimation functions

These functions allow to estimate the mte and mtr, and their confidence intervals, based on coefficients estimated from the main model in the main function. More details can be found in Bruneel-Zupanc (2024).
Different formulas must be applied depending on whether the treatment is homogenous or heterogenous.

Usage

mtr_est_poly(d, data, seq_u,
                    bwd = NULL, bw_y = NULL, bw_method = "plug-in",
                    pol_degree1, pol_degree2,
                    var_outcome, var_treatment, var_w0, var_w1, var_covariates,
                    print_progress)

mtr_fun_poly(ref_indiv, eval_v, est0, est1, k0, k1, se_type)

lpoly_regress(x, y, bw=NULL, bw_method="plug-in", degree, drv)

mtr_coeff(coeff, vcov, var_cov_2nd, est_method="sieve")

mtr_est(coeff, vcov, names_var, df)

mtr_predict_sieve(coeff, vcov, ref_indiv, var_treatment, var_cov_2nd, pol_degree, seq_u, t_value, est_method, se_type)

Arguments

seq_u

Sequence of v at which to compute the prediction for kd(v).

bw_method

Method to compute the bandwidth (if bandwdith is NULL)

ref_indiv

Newdata (reference individuals) at which to compute the predictions.

x

Vector of x values

y

Vector of y values

bw

Pre-specified bandwidth

degree

Degree of the polynomial: recommended to set to drv + 1

drv

Derivative order of the function to be estimated.

coeff

Vector of (stacked) coefficients for mtr0, mtr1.

vcov

Covariance matrix of these coefficients.

est_method

Either "sieve" or "homogenous".

names_var

Names of the variables corresponding to the coefficients.

df

Degrees of freedom for the p-values.

var_cov_2nd

Names of the covariates and semi-IVs

Value

Returns the raw (stacked regression) coefficients and covariance matrix corresponding to mtr0, mtr1 and mte function.

Returns the mtr0, mtr1 and mte estimates tables with their standard errors and p-values. Also exports the corresponding estimates and vcov matrices.

Returns mtr0, mtr1, and mte estimates with confidence intervals for the specified ref_indiv.