Simulate data from the Generalized Roy Model with semi-IVs

This function simulates data from the Generalized Roy Model with semi-IVs.
This is the simulation specified in Bruneel-Zupanc (2024).

Usage

simul_data(N, model_type="heterogenous",
           param_y0, param_y1, param_p, param_Z, param_genX, param_error)

Arguments

N

Number of observations

model_type

Type of model: "heterogenous" or "homogenous"

param_y0

Parameters for Y0 = \((\delta_0, \beta_0, \beta_{0X_1}, \beta_{0X_2})\).
i.e., intercept, effects on w0, \(X_1\), \(X_2\) on \(Y_0\).

param_y1

Parameters for Y1: \((\delta_1, \beta_1, \beta_{1X_1}, \beta_{1X_2})\).
i.e., intercept, effects w1, \(X_1\), \(X_2\) on \(Y_1\).

param_p

Parameters for the selection: \((\alpha, \alpha_0, \alpha_1, \alpha_2, \alpha_{X_1}, \alpha_{X_2})\)
i.e., intercept and effects of w0, w1, w0w1, Xbinary, Xcontinuous on the latent utility.

param_Z

Parameters for the simulation of the semi-IVs:
\((\mu_{W0,x_1=0}, \mu_{W1,x_1=0}, \mu_{W0,x_1=1}, \mu_{W1,x_1=1}, \sigma^2_{W0}, \sigma^2_{W1}, \sigma_{W0W1})\).

param_genX

Parameters for the covariates: \((p_{X_1}, \mu_{X_2}, \sigma^2_{X_2})\).

param_error

Parameters for the error terms: depends on model_type:
if heterogenous: \((\sigma^2_{U0}, \sigma^2_{U1}, \sigma_{U0U1}, \sigma^2_{\text{cost}})\)
if homogenous: \((\sigma^2_{U}, \sigma^2_{V}, \sigma_{UV})\)

Value

A data frame with the following columns:

y

The observed outcome.

d

The treatment.

w0, w1

The semi-IVs entering only D=0 and D=1.

Xbinary, Xcontinuous

Two covariates, one binary and one continuous.

y0, y1

The unobserved potential outcomes.

P

The unobserved true treatment probability.

latent, V, Ud, U0, U1

The unobserved shocks V. Ud is the normalized V ranks. U0 and U1 are the outcome shocks. latent gives the latent utility term in the selection equation.

The Generalized Roy model.

This function simulates a generalized Roy model as described in Bruneel-Zupanc (2024).

Potential Outcomes. The potential outcomes (e.g., earnings) are given by: $$ Y_0 = \delta_{0} + \beta_{0} W_0 + X \beta_{0X} + U_0, $$ $$ Y_1 = \delta_{1} + \beta_{1} W_1 + X \beta_{1X} + U_1, $$ where \(W_0, W_1\) are the observed semi-IVs excluded from \(Y_1\) and \(Y_0\) respectively, \(X=(X_1, X_2)\) is a vector of a binary (\(X_1\), e.g., location) and a continuous (\(X_2\), e.g., education of the parents) observable covariates, and \(U_0, U_1\) are unobservable errors.

Selection problem. We only observe the outcome $$ Y = (1-D) Y_0 + D Y_1. $$ where \(D\) represents the (binary) treatment decision (e.g., education choice), given by the following selection rule: $$ \begin{aligned} D^* &= g(W_0, W_1, X) - V \\ &= - ( \alpha + \alpha_0 W_0 + \alpha_1 W_1 + \alpha_2 W_0 \times W_1 + \alpha_{X_1} X_1 + \alpha_{X_2} X_2) - V, \\ \text{ with } \quad D &= \mathbb{I}(D^* > 0), \end{aligned} $$ where \(V\) is the main unobservable probability shock, and the higher \(V\), the more likely one is to be treated. Note that we normalize \(U_D=F_{V}(V)\) to get the normalized probability shock \(U_D \sim \mathcal{U}(0, 1)\). \(U_D\) can be interpreted as unobserved resistance to treatment, The closer \(U_D\) is to 0, the more likely the individual is to be treated.

This specification yields that the probability of treatment is given by: $$ \textrm{Pr}(D=1 | W_0, W_1, X) = \textrm{Pr}(V < g(W_0, W_1, X)). $$ Thus, ceteris paribus, the lower \(g\), the higher the probability of treatment.

Unobservables specification. The specification of the unobservable depends on the model type.

For the general heterogenous treatment effect model, we have: #' $$ \begin{pmatrix} U_0 \\ U_1 \end{pmatrix} \sim \mathcal{N}\left( \begin{pmatrix} 0 \\ 0 \end{pmatrix}, \begin{pmatrix} \sigma^2_{U0} & \sigma_{U0U1} \\ \sigma_{U0U1} & \sigma^2_{U1} \end{pmatrix} \right), $$ $$ C \sim \mathcal{N}(\mu_{\text{cost}}, \sigma^2_{\text{cost}}), $$ $$ V = -(U_1 - U_0 - C). $$

For the homogeneous treatment effect model: $$ \begin{pmatrix} U \\ V \end{pmatrix} \sim \mathcal{N}\left( \begin{pmatrix} 0 \\ \mu_{V} \end{pmatrix}, \begin{pmatrix} \sigma^2_{U} & \sigma_{UV} \\ \sigma_{UV} & \sigma^2_{V} \end{pmatrix} \right), $$ $$ U_0 = U_1 = U. $$

In both cases, \(V\) is normally distributed, such that the selection equation is a probit model.

Covariates and semi-IVs specification. The covariates are generated by $$ X_1 \sim \text{Bernoulli}(p_{X_1}) \text{ and } X_1 \sim \mathcal{N}(\mu_{X_2}, \sigma^2_{X_2}). $$

The semi-IVs are \(X_1\)-specific and are given by: $$ \begin{pmatrix} W_0 \\ W_1 \end{pmatrix} \sim \mathcal{N}\left( \begin{pmatrix} \mu_{W0,x_1} \\ \mu_{W1,x_1} \end{pmatrix}, \begin{pmatrix} \sigma^2_{W0} & \sigma_{W0W1} \\ \sigma_{W0W1} & \sigma^2_{W1} \end{pmatrix} \right), $$ where the means \(\mu_{W0,x_1}\) and \(\mu_{W1,x_1}\) depend on the binary covariate \(X_1=x_1\).

Simulating standard IV MTE models.

This function can be used to model problem with IVs used to estimate Marginal Treatment Effects, by setting the effect of the semi-IVs on their respective outcomes to zero. The coefficients can be adjusted to mimic the Roy models of Heckman, Urzua and Vytlacil (2006), or Heckman and Vytlacil (2007). Small adjustments inside the function allow to mimic the simulation of Andresen (2018) (mtefe in Stata), but with only 2 regions (state).

References

Bruneel-Zupanc, C. (2023). Don't (fully) exclude me, it's not necessary! Identification with semi-IVs. arXiv preprint arXiv:2303.12667.

Andresen, M. E. (2018). Exploring marginal treatment effects: Flexible estimation using Stata. The Stata Journal, 18(1), 118-158.

Heckman, J. J., Urzua, S., & Vytlacil, E. (2006). Understanding instrumental variables in models with essential heterogeneity. The Review of Economics and Statistics, 88(3), 389-432.

Heckman, J. J., & Vytlacil, E. J. (2007). Econometric evaluation of social programs, part II: Using the marginal treatment effect to organize alternative econometric estimators to evaluate social programs, and to forecast their effects in new environments. Handbook of econometrics, 6, 4875-5143.

Examples

N = 10000; set.seed(12345)

# Example 1: Heterogenous Treatment Effects.
model_type = "heterogenous"
param_error = c(1, 1, 0.6, 0.5) # var_u0, var_u1, cov_u0u1, var_cost
param_Z = c(0, 0, 0, 0, 1.5, 1.5, 0.9)
# meanW0 Xbinary0, meanW1 Xbinary0, meanW0 Xbinary1, meanW1 Xbinary1, varW0, varW1, covW0W1
param_p = c(0, -0.7, 0.7, 0, 0, 0) # constant, W0, W1, W0xW1, Xbinary, Xcontinuous
param_y0 = c(3.2, 0.8, 0, 0) # intercept, W0, Xbinary, Xcontinuous;
param_y1 = c(3.2+0.4, 0.5, 0, 0) # the +0.4 = ATE; W1, Xbinary, Xcontinuous;
param_genX = c(0.4, 0, 2)

data = simul_data(N, model_type, param_y0, param_y1, param_p, param_Z, param_genX, param_error)

# Example 2: Homogenous Treatment Effects (constant MTE)
model_type = "homogenous"
param_error = c(1, 1.5, -0.6) # var_u, var_v, cov_uv
param_Z = c(0, 0, 0, 0, 1.5, 1.5, 0.9)
param_p = c(0, -0.5, 0.5, 0, 0, 0) # the constant <=> mean_V
param_y0 = c(3.2, 0.8, 0, 0)
param_y1 = c(3.2+0.4, 0.5, 0, 0)
param_genX = c(0.4, 0, 2)

data1 = simul_data(N, model_type, param_y0, param_y1, param_p, param_Z, param_genX, param_error)

# Set the effects of w1 or w0 on its outcome to zero if want a valid IV, e.g.,
# param_y1 = c(3.2+0.4, 0, 0, 0) # w1 is a valid IV
# or: param_y0 = c(3.2, 0, 0, 0) # w0 is a valid IV